Refactor expose Counters under /htmx/counters
