+++ src/components/counter.js
@@ -11,11 +11,11 @@
 export default ({
   van, id, init = 0, buttonStyle = "👍👎",
 }) => {
-  const {button, div} = van.tags
+  const { button, div } = van.tags
 
   const [up, down] = [...van.val(buttonStyle)]
   const counter = van.state(init)
-  console.log("counter", counter);
+
   return div({...(id ? {id} : {}), "data-counter": counter},
     "❤️ ", counter, " ",
     button({onclick: () => {
